chatbot_app/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
chromadb==1.0.15
openai==1.98.0
python-dotenv==1.0.1
pandas==2.2.2
instructor==1.10.0
numpy<2.0
mistralai==1.9.3
python-multipart==0.0.9
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
httpx==0.28.1
aiofiles==24.1.0
pydub==0.25.1
ffmpeg-python==0.2.0
google-generativeai==0.7.1